Move trait definitions from lib.rs to interface.rs
